Sistema/motor/requirements.txt
python-docx
colorama
jinja2
google-generativeai
anthropic
pdfplumber
pypdf
docling
rapidocr-onnxruntime
